REFACTOR : Todo조회시 학습시간도 내려주도록 수정, 과제제출/완료 API 통합

+++ src/main/java/com/momen/application/planner/dto/StudyDailyStatsResponse.java
@@ -16,5 +16,5 @@ public class StudyDailyStatsResponse {
     private int total;
     private int completed;
     private int remaining;
-    private double completionRatePercent;
+    private int completionRatePercent;
 }
